refactor: seperate Date component
