+++ src/components/recordDashboard/RecordCardLayout.tsx
@@ -1,57 +1,13 @@
 import { useState, useContext } from "react";
 import moment from "moment";
-import {
-  IconChevronLeft,
-  IconChevronRight,
-  IconCirclePlus,
-} from "@tabler/icons-react";
+import { IconChevronUp, IconCirclePlus } from "@tabler/icons-react";
 import { DateContext } from "@/pages/record_dashboard";
+import Date from "./recordCard/Date";
 import { originalData } from "@/common/lib/test/eventData";
 
 const RecordCardLayout = () => {
   const { selectedDate } = useContext(DateContext);
 
-  const Date = () => {
-    const [currentTime, setCurrentTime] = useState(moment(selectedDate));
-
-    const handlePreviousDay = () => {
-      setCurrentTime(currentTime.clone().subtract(1, "day"));
-    };
-    const handleNextDay = () => {
-      setCurrentTime(currentTime.clone().add(1, "day"));
-    };
-
-    return (
-      <div className="flex justify-center items-center gap-x-4">
-        <IconChevronLeft
-          size={24}
-          color={"#808080"}
-          className="hover:cursor-pointer"
-          onClick={handlePreviousDay}
-        />
-        <ul className="flex gap-x-2 text-2xl justify-center font-medium max-w-[124px] w-full">
-          <ol className="flex gap-x-1">
-            <li className="font-['futura'] text-center w-[30px]">
-              {currentTime.format("M")}
-            </li>
-            <li>月</li>
-          </ol>
-          <ol className="flex gap-x-1">
-            <li className="font-['futura'] text-center w-[30px]">
-              {currentTime.format("D")}
-            </li>
-            <li>日</li>
-          </ol>
-        </ul>
-        <IconChevronRight
-          size={24}
-          color={"#808080"}
-          className="hover:cursor-pointer"
-          onClick={handleNextDay}
-        />
-      </div>
-    );
-  };
   const AddRecordBtn = () => {
     const [isShown, setIsShown] = useState(false);
     const ShowRecordType = () => {
